+++ monkey/monkey_island/cc/services/utils/network_utils.py
@@ -1,10 +1,8 @@
 import array
-import collections
 import ipaddress
 import socket
 import struct
 import sys
-from typing import List
 
 from netifaces import AF_INET, ifaddresses, interfaces
 from ring import lru
@@ -51,11 +49,6 @@ else:
                     # name of interface is (namestr[i:i+16].split('\0', 1)[0]
         finally:
             return result
-
-
-def is_local_ips(ips: List) -> bool:
-    filtered_local_ips = [ip for ip in local_ip_addresses() if not ip.startswith("169.254")]
-    return collections.Counter(ips) == collections.Counter(filtered_local_ips)
 
 
 # The local IP addresses list should not change often. Therefore, we can cache the result and